refactor: move ParseDateTag out of internal/types (#781)
internal/types holds wire-compatible core data structures. Tag-parsing is
operational logic; it was placed in types/ to dodge an import cycle.
Move it to a new internal/parse package and update call sites in
internal/mcp and internal/db. Tests follow the function.

Removes the now-unused strings import from internal/types.
