+++ internal/parse/date_tag_test.go
@@ -1,15 +1,13 @@
-package mcp
+package parse
 
 import (
 	"testing"
 	"time"
-
-	"github.com/petersimmons1972/engram/internal/types"
 )
 
 func TestParseDateTag_Valid(t *testing.T) {
 	tags := []string{"lme", "sid:abc123", "date:2023-06-15"}
-	got := types.ParseDateTag(tags)
+	got := ParseDateTag(tags)
 	if got == nil {
 		t.Fatal("parseDateTag returned nil, want 2023-06-15")
 	}
@@ -21,30 +19,30 @@ func TestParseDateTag_Valid(t *testing.T) {
 
 func TestParseDateTag_Invalid(t *testing.T) {
 	tags := []string{"lme", "date:not-a-date"}
-	got := types.ParseDateTag(tags)
+	got := ParseDateTag(tags)
 	if got != nil {
-		t.Errorf("types.ParseDateTag(%q) = %v, want nil for invalid date", tags, got)
+		t.Errorf("ParseDateTag(%q) = %v, want nil for invalid date", tags, got)
 	}
 }
 
 func TestParseDateTag_NonePresent(t *testing.T) {
 	tags := []string{"lme", "sid:abc", "project:foo"}
-	got := types.ParseDateTag(tags)
+	got := ParseDateTag(tags)
 	if got != nil {
 		t.Errorf("parseDateTag with no date: tag = %v, want nil", got)
 	}
 }
 
 func TestParseDateTag_EmptyTags(t *testing.T) {
-	got := types.ParseDateTag(nil)
+	got := ParseDateTag(nil)
 	if got != nil {
-		t.Errorf("types.ParseDateTag(nil) = %v, want nil", got)
+		t.Errorf("ParseDateTag(nil) = %v, want nil", got)
 	}
 }
 
 func TestParseDateTag_FirstWins(t *testing.T) {
 	tags := []string{"date:2022-01-01", "date:2024-12-31"}
-	got := types.ParseDateTag(tags)
+	got := ParseDateTag(tags)
 	if got == nil {
 		t.Fatal("parseDateTag returned nil")
 	}
@@ -67,13 +65,13 @@ func TestParseDateTag_LongMemEvalFormat(t *testing.T) {
 		{"date:2024/01/01 (Mon) 09:00", time.Date(2024, 1, 1, 0, 0, 0, 0, time.UTC)},
 	}
 	for _, tc := range cases {
-		got := types.ParseDateTag([]string{tc.tag})
+		got := ParseDateTag([]string{tc.tag})
 		if got == nil {
-			t.Errorf("types.ParseDateTag(%q) = nil, want %v", tc.tag, tc.want)
+			t.Errorf("ParseDateTag(%q) = nil, want %v", tc.tag, tc.want)
 			continue
 		}
 		if !got.Equal(tc.want) {
-			t.Errorf("types.ParseDateTag(%q) = %v, want %v", tc.tag, *got, tc.want)
+			t.Errorf("ParseDateTag(%q) = %v, want %v", tc.tag, *got, tc.want)
 		}
 	}
 }
